+++ src/djlint/formatter/condense.py
@@ -1,4 +1,8 @@
-"""djLint Attempt to logically shrink html."""
+"""Condense HTML.
+
+1. Put short html tags back on one line
+2. Put short templage tags back on one line
+"""
 
 from functools import partial
 
@@ -8,7 +12,7 @@ from ..helpers import inside_ignored_block
 from ..settings import Config
 
 
-def compress_html(html: str, config: Config) -> str:
+def condense_html(html: str, config: Config) -> str:
     """Compress back tags that do not need to be expanded."""
     # put empty tags on one line
 
@@ -23,13 +27,24 @@ def compress_html(html: str, config: Config) -> str:
 
     html = re.sub(re.compile(r"^[ \t]*(.*?)[\n \t]*$", re.M), func, html)
 
+    def condense_line(config: Config, match: re.Match) -> str:
+        """Put contents on a single line if below max line length."""
+        if (
+            len(match.group(1) + match.group(3) + match.group(4))
+            < config.max_line_length
+        ):
+            return match.group(1) + match.group(3) + match.group(4)
+        return match.group()
+
+    func = partial(condense_line, config)
+
     # put short single line tags on one line
     html = re.sub(
         re.compile(
-            fr"(<({config.optional_single_line_html_tags})(?:\s(?:(?:{{%[^(?:%}}]*?%}})|(?:{{{{[^(?:}}}})]*?}}}})|[^<>])*)?>)\s*([^<\n]{{,80}})\s*?(</(\2)>)",
+            fr"(<({config.optional_single_line_html_tags})\b(?:\"[^\"]*\"|'[^']*'|{{[^}}]*}}|[^'\">{{}}])*>)\s*([^<\n]*?)\s*?(</(\2)>)",
             re.IGNORECASE | re.MULTILINE | re.DOTALL | re.VERBOSE,
         ),
-        r"\1\3\4",
+        func,
         html,
         re.IGNORECASE | re.MULTILINE | re.DOTALL,
     )
@@ -37,10 +52,10 @@ def compress_html(html: str, config: Config) -> str:
     # put short template tags back on one line
     html = re.sub(
         re.compile(
-            rf"({{%-?[ ]*?({config.optional_single_line_template_tags})[^\n(?:%}})]{{,30}}%}})\s*([^%\n]{{,50}})\s*?({{%-?[ ]+?end(\2)[ ]*?%}})",
+            rf"({{%-?[ ]*?({config.optional_single_line_template_tags})[^\n(?:%}})]*?%}})\s*([^%\n]*?)\s*?({{%-?[ ]+?end(\2)[ ]*?%}})",
             flags=re.IGNORECASE | re.MULTILINE | re.VERBOSE,
         ),
-        r"\1\3\4",
+        func,
         html,
     )