+++ src/djlint/formatter/expand.py
@@ -1,25 +1,15 @@
-"""djLint expand out html code."""
+"""Expand html.
+
+1. put html tags on individual lines, if needed.
+2. puth template tags on individual lines, if needed.
+"""
+
 from functools import partial
 
 import regex as re
 
 from ..helpers import inside_ignored_block
 from ..settings import Config
-
-
-def _flatten_attributes(match: re.Match) -> str:
-    """Flatten multiline attributes back to one line.
-
-    Skip when attribute is ignored.
-    Attribute name can be in group one or group 2.
-    for now, skipping if they are anywhere
-    """
-    # pylint: disable=C0209
-    return "{} {}{}".format(
-        match.group(1),
-        " ".join(x.strip() for x in match.group(2).strip().splitlines()),
-        match.group(3),
-    )
 
 
 def expand_html(html: str, config: Config) -> str:
@@ -37,25 +27,16 @@ def expand_html(html: str, config: Config) -> str:
 
     html_tags = config.break_html_tags
 
-    # put attributes on one line
-    html = re.sub(
-        re.compile(
-            fr"(<(?:{config.indent_html_tags})\b)((?:\s(?:(?:{{%[^(?:%}}]*?%}})|(?:{{{{[^(?:}}}})]*?}}}})|[^<>/])*)+?)(/?>)",
-            flags=re.IGNORECASE | re.MULTILINE | re.VERBOSE,
-        ),
-        _flatten_attributes,
-        html,
-    )
-
     add_left = partial(add_html_line, "\n%s")
     add_right = partial(add_html_line, "%s\n")
 
     break_char = config.break_before
 
     # html tags - break before
+
     html = re.sub(
         re.compile(
-            fr"{break_char}\K(</?(?:{html_tags})\b(\"[^\"]*\"|'[^']*'|[^'\">])*>)",
+            fr"{break_char}\K(</?(?:{html_tags})\b(\"[^\"]*\"|'[^']*'|{{[^}}]*}}|[^'\">{{}}])*>)",
             flags=re.IGNORECASE | re.VERBOSE,
         ),
         add_left,
@@ -65,7 +46,7 @@ def expand_html(html: str, config: Config) -> str:
     # html tags - break after
     html = re.sub(
         re.compile(
-            fr"(</?(?:{html_tags})\b(\"[^\"]*\"|'[^']*'|[^'\">])*>)(?=[^\n])",
+            fr"(</?(?:{html_tags})\b(\"[^\"]*\"|'[^']*'|{{[^}}]*}}|[^'\">{{}}])*>)(?=[^\n])",
             flags=re.IGNORECASE | re.VERBOSE,
         ),
         add_right,
@@ -73,7 +54,6 @@ def expand_html(html: str, config: Config) -> str:
     )
 
     # template tag breaks
-
     def should_i_move_template_tag(out_format: str, match: re.Match) -> str:
         # ensure template tag is not inside an html tag