+++ tests/test_html.py
@@ -5,7 +5,7 @@ run::
    pytest tests/test_html.py --cov=src/djlint --cov-branch \
           --cov-report xml:coverage.xml --cov-report term-missing
 
-   pytest tests/test_html.py::test_self_closing_tags --cov=src/djlint --cov-branch \
+   pytest tests/test_html.py::test_ignored_attributes --cov=src/djlint --cov-branch \
           --cov-report xml:coverage.xml --cov-report term-missing
 
 
@@ -251,7 +251,7 @@ def test_ignored_attributes(runner: CliRunner, tmp_file: TextIO) -> None:
      id="somthing_meaning_less_is_here"
      required
      checked="checked"
-     json-data='{"menu":{"header":"SVG Viewer","items":[{"id":"Open"}]}}'></div>
+     json-data='{"menu":{"header":"SVG Viewer","items":[{"id":"Open"}]}}'>\n</div>
 """
     )
 
@@ -302,9 +302,21 @@ def test_ignored_block(runner: CliRunner, tmp_file: TextIO) -> None:
     output = reformat(
         tmp_file,
         runner,
-        b"""{% djlint:off %}
-    <div><p><span></span></p></div>
-{% djlint:on %}
+        b"""<!-- djlint:off -->
+<div><p><span></span></p></div>
+<!-- djlint:on -->
+{# djlint:off #}
+<div><p><span></span></p></div>
+{# djlint:on #}
+{% comment %} djlint:off {% endcomment %}
+<div><p><span></span></p></div>
+{% comment %} djlint:on {% endcomment %}
+{{ /* djlint:off */ }}
+<div><p><span></span></p></div>
+{{ /* djlint:on */ }}
+{{!-- djlint:off --}}
+<div><p><span></span></p></div>
+{{!-- djlint:on --}}
 """,
     )
 
@@ -312,9 +324,21 @@ def test_ignored_block(runner: CliRunner, tmp_file: TextIO) -> None:
 
     assert (
         output["text"]
-        == """{% djlint:off %}
-    <div><p><span></span></p></div>
-{% djlint:on %}
+        == """<!-- djlint:off -->
+<div><p><span></span></p></div>
+<!-- djlint:on -->
+{# djlint:off #}
+<div><p><span></span></p></div>
+{# djlint:on #}
+{% comment %} djlint:off {% endcomment %}
+<div><p><span></span></p></div>
+{% comment %} djlint:on {% endcomment %}
+{{ /* djlint:off */ }}
+<div><p><span></span></p></div>
+{{ /* djlint:on */ }}
+{{!-- djlint:off --}}
+<div><p><span></span></p></div>
+{{!-- djlint:on --}}
 """
     )
 
@@ -331,11 +355,11 @@ def test_ignored_block(runner: CliRunner, tmp_file: TextIO) -> None:
     assert output["exit_code"] == 0
 
     assert (
-        output["text"]
-        == """<script>
+        """<script>
     <div><p><span></span></p></div>
 </script>
 """
+        in output["text"]
     )
 
     # check inline script includes
@@ -351,7 +375,7 @@ def test_ignored_block(runner: CliRunner, tmp_file: TextIO) -> None:
 </html>
 """,
     )
-
+    print(output["text"])
     assert output["exit_code"] == 0