+++ tests/test_linter.py
@@ -7,7 +7,7 @@ run::
 
    # for a single test
 
-   pytest tests/test_linter.py::test_H026 --cov=src/djlint --cov-branch \
+   pytest tests/test_linter.py::test_DJ018 --cov=src/djlint --cov-branch \
          --cov-report xml:coverage.xml --cov-report term-missing
 
 """